move bin/*.py to tron/bin and symlink back, fix relevant tests
